hw4/requirements.txt
fastapi==0.118.2
psycopg==3.2.10
psycopg-binary==3.2.10
uvicorn==0.37.0
faker==37.11.0
pyjwt==2.10.1
redis==6.4.0